Add global constants for CSV file paths
For convenience, I defined constants in the data_preparation.py
file that point to strings of the file paths, so from within the
tools directory the files should be easily accessible from outside
the data/ directory. This can be changed in the future depending
on the layout of the repository.