Add clean_rt_movie_info function
This returns a cleaned DataFrame from the rt.movie_info.tsv file.
Also, default paths were set for this and clean_rt_reviews using
defined constants in the data_preparation.py file. The dropna
parameter when True will drop all rows with NaN values in the
subset column(s), or in any column if not specified.