Set global path constants relative to main directory
Previously, these were relative to the tools package (i.e.
"../data/<filename>", and not "./data/<filename>"). Now,
there is no need to modify the path or change the working
directory when in the main directory. When working in a
file within a subdirectory, for now you should change
the working directory to microsoft_movies